Fixed Saving ProblemDeleted all Migrations and deleted all tables then re-did Migration from scratch
Fixed Saving ProblemDeleted all Migrations and deleted all tables then
re-did Migration from scratch

+++ GigHub/Models/Genre.cs
@@ -5,7 +5,8 @@ namespace GigHub.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
+        
+        public byte Id { get; set; }
 
         [Required]
         [StringLength(255)]
